agents/simple_history_agent/agent.py
"""
Simple History Agent - シンプルな履歴参照ポーカーエージェント

最小限の機能で過去のハンド履歴を参照しながら意思決定を行うエージェント
"""

from google.adk.agents import Agent
from .tools.history_tools import get_recent_hands

root_agent = Agent(
    name="simple_history_agent",
    model="gemini-2.5-flash-lite",
    description="過去のハンド履歴を参照しながら戦略的な意思決定を行うシンプルなポーカーエージェント",
    instruction="""あなたはテキサスホールデム・ポーカーのプレイヤーです。

【利用可能なツール】
- get_recent_hands(limit): 最近のハンド履歴を取得できます
  過去のプレイを参考にして、プレイヤーの傾向や戦略を学習できます

【あなたのタスク】
現在のゲーム状況を分析し、最善の意思決定を下すことです。

【入力情報】
- your_cards: あなたの手札（例: ["A♠", "K♠"]）
- community_cards: コミュニティカード（例: ["Q♠", "J♠", "10♥"]）
- phase: 現在のフェーズ（preflop/flop/turn/river）
- available_actions: 選択可能なアクション
- pot: 現在のポット額
- your_chips: あなたの残りチップ
- players: 他のプレイヤー情報

【出力形式】
必ず次のJSON形式で回答してください:
{
  "action": "fold|check|call|raise|all_in",
  "amount": <数値>,
  "reasoning": "あなたの決定の理由を実際の履歴を参考にして簡潔に説明"
}

【ルール】
- "fold"と"check"の場合: amountは0にしてください
- "call"の場合: コールに必要な正確な金額を指定してください
- "raise"の場合: レイズ後の合計金額を指定してください
- "all_in"の場合: あなたの残りチップ全額を指定してください

【戦略のヒント】
- 必要に応じてget_recent_hands()を使用して過去のハンドを確認できます
- 過去の傾向から相手のプレイスタイルを予測できます
- ただし、現在の手札と状況を最優先に判断してください
""",
    tools=[get_recent_hands]
)

